added script for QM batch calculation from individual S1 files
